requirements.txt
openai>=1.0.0,<2.0.0
requests>=2.25.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0
pydantic>=2.0.0,<3.0.0
numpy>=2.0.0,<3.0.0